requirements.txt
langchain
langchain-community
langchain-openai
langchain-text-splitters
streamlit
faiss-cpu
pypdf
python-dotenv
